[roll] Roll fuchsia [fvm] Remove redundant DEVICE_ADD_INVISIBLE, migrate RemoveDeprecated.
This driver was already migrated to the new DdkInit hook in fxr/407130,
but forgot to remove this arg, which now has no effect.

DdkAsyncRemove is the same as DdkRemoveDeprecated but also calls
the unbind hook. Since the VPartition device implements a trivial
unbind, this has no change in behavior. DdkAsyncRemove does however
assert when trying to remove a device when zxdev_ == nullptr,
so have added a check.

BUG=37831
BUG=43261

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413354
Original-Revision: f0feda570110d4c666b0b9a735381abd76fc11cb
GitOrigin-RevId: b4982d52a67e67f49aa9308e01a03d850203e94c
Change-Id: I4eeac138b7595e6bb65a64bdaf5a13de1f69b607