[roll] Roll fuchsia [bt][fidl] Create helper function to convert ServiceDefinition to ServiceRecord
Move code that converts a ServiceDefinition to a ServiceRecord from
profile_server.Advertise() to separate functions in the FIDL helper
library so that they can be accessed outside the profile server.

This is useful to allow for the creation and use of a SDP search results
field in fuchsia.bluetooth.test.BredrPeerParameters.

Test: fx test bt-host-unittests -- --gtest_filter="SDP_ServerTest",
FidlHelperTest.ServiceDefinitionToServiceRecord
Original-Bug: 57370

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412062
Original-Revision: c121a5b6340a4596d6c13286f2e57efa81fddc06
GitOrigin-RevId: 9a4158578b64b940b737214c92ee2f92ac82cd4a
Change-Id: I22754b25a6d7aede87d0ce6a2cd1ca19f4257662