[roll] Roll fuchsia [fidl][llcpp] merge Reply and ReplyWithStatus
Now Reply returns a status (which can be ignored).

Fixes the code which use it.

Also remove the version with a DecodedMessage argument which is not used
and not tested. This version can be reintroduced later with another type
(DecodedMessage will be removed).

That spares 75484 bytes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413020
Original-Revision: 34bd145784e2e3b8fb305a8d18fe28ba32db2746
GitOrigin-RevId: 244b221ca95159232324698fc7e367c1da959b78
Change-Id: Iffeae4a0fe6b975f3ceed9fd319e301a54e1097c