[roll] Roll fuchsia [zircon][debugger] fix logic in mxcsr_is_valid.
The function `mxcsr_is_valid` had a misleading name: it would return
true if the given value for an mxcsr was invalid, and false if it was
valid.

The code using it used it correctly, a more precise name would have been
`mxcsr_is_invalid`.

This CL modifies the code in mxcsr_is_valid so that it returns true if
the given value for an mxcsr is valid, and false otherwise.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405487
Original-Revision: 610085d933a8524e6adf52224c1c349280ff27df
GitOrigin-RevId: 3202df8b940b630e70f3a7cd10737472995591b8
Change-Id: I45a1d06b82774b6c703055a702dbb4ac3c403646