[roll] Roll fuchsia [net-cli] Add route table add/del
This change adds interfaces to net-cli which afford adding and deleting
routes from the route table. This is part of an effort to bring
interfaces that currently only exist in ifconfig into net-cli, with the
aim of eventually deprecating and removing the former.

Original-Fixed: 56758

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414662
Original-Revision: 999268ea6fe0100ea1fc060796070e98fc45da32
GitOrigin-RevId: cefe12d5a3d09deb8973e275bf6ec0f6857d4744
Change-Id: I0a1038d675acef78cd3ff339707106a6bb8b8974