[roll] Roll fuchsia [fx] Remove error message  in fx-env.sh for unknown args
For some WFH workflows, we need to execute scripts with $PATH initialzed
using the $USER's environment. This is done by using the shebang for
those scripts of "#!/bin/bash -i" which creates an interactive shell to
execute the script.

Some developers have added `source fx-env.sh` to their .bashrc files
as we suggest in the getting started guide. This script is parsing
the command line arguments used to start the shell, and incorrectly
assumes that any argurments are from the source command, and not the
command line of bash.

This removes the error message for unknown arguments and they are
ignored.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414742
Original-Revision: 7e833d4c45f5e929d6614a1fea52eabf0afc0680
GitOrigin-RevId: b820735b46fc440e60bf20b5e090d3041aba93ff
Change-Id: I6f2857c15b288e973aa43ab239718704bb53b9c8