[roll] Roll fuchsia [bt][sm] Make SM dependency-injectable
Introduce an abstract class with the production public interface of
sm::SecurityManager and only basic getters/setters a la l2cap::Channel,
moving the production implementation to a separate SecurityManagerImpl.
This is in preparation for a follow-on change which introduces an SM
test double for use in GAP unit tests.

Original-Bug: 886
Test: Pure refactor, run `$ fx test bt-host-unittests` to verify nothing
      broke
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414455
Original-Revision: 179f0dd14f1f79127b1756e84b6e7097757eacd9
GitOrigin-RevId: 6d1b07ab2813054de6058054717dba588c736e3c
Change-Id: I261de09a7712123a9fdf7174895134622b447cf7