[roll] Roll fuchsia [blobfs] remove `blob_header_next` from superblock
This field has never been used for ~3 years, and there do not appear to
be any plans to use them. We verify the field is zeroed when mounting
the filesystem.

This change does not modify the value in the field, and continues to
ensure the field is zero on mount, so it is a NOP.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404619
Original-Revision: 96f46fee5f9215e6c2176be733d91264f7b3e2af
GitOrigin-RevId: bf64525ea328a2b3342ac2608bbd1468982bdef9
Change-Id: I76089b1638c309a265cee1f165d8270dde83bb2a