[roll] Roll fuchsia [wlan] Don't log no-op disconnect_ctx in Inspect
Only log the `disconnect_ctx` if an operation had an effect of moving from
non-idle state to idle state. This is so that the client that consumes
`disconnect_ctx` does not log a disconnect event when it's effectively no-op.

Original-Bug: b/154549756, b/161450584
Test: Logging change only. Manually verified Inspect log.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412594
Original-Revision: bf7485a6f795dba725fcf4d5afd65621114a3b8c
GitOrigin-RevId: 3696b11df1711643dd6e673a79028016c46dc364
Change-Id: Ia96f7219ff0db125eb8ede3d0285e3568d3510b8