[roll] Roll fuchsia [bt][bt-gap] Persist discovery state over active host adapter change
By decoupling the state of Discovery into a dispatcher session and host
session, we are able to provide clients with a dispatcher session while
managing the host sessions at the dispatcher level, allowing us to
seemlessly change the host on which Discovery is ocurring without
perturbing the client's discovery session.

A potential awkwardness with my current implementation is not handling
the acquisition of `HostDiscoverySession` in the `impl DiscoveryState`
methods; as it is, calling `discover_on_active_host` occurs outside the
`DiscoveryState` methods and `HostDiscoverySession` is passed in as an
argument as necessary. It is possible to subsume this functionality
into the `DiscoveryState` methods. This would require passing in a
reference to the `HostDispatcher` to the methods, and potentially
rewriting `discover_on_active_host` to return `impl Future<>` to
avoid borrowing issues.  The advantage to this is that interfacing with
`HostDiscoverySession` will be managed entirely within the
DiscoveryState API.

Original-Bug: 640
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410007
Original-Revision: 78ebbced84b2f676143831cdd6ef5ca1b2338ae4
GitOrigin-RevId: 75cec8f88617d265d12d770d332308b4bb0bd361
Change-Id: Ic8a7c3e7b414efbf7ba00f9a2d76b20e1101a0e0