[roll] Roll cobalt Deprecate the report_definition's output_location and remove validation of it.
The data for this field is now being stored in an intenral Cobalt
metadata location, and so it doesn't need to be in the registry.
Subsequent changes will remove the setting of this field from the
registry YAML files, and then eventually remove the field
completely.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/414681
Original-Revision: 2815a7d5f49abb70174b40046942c0ccac1a8f9e
GitOrigin-RevId: 5e7461add9f6974465696188328128a26de2e73f
Change-Id: I0fbc4e75ec73aee0a3ee58355819853ec5dd03aa