[roll] Roll fuchsia [memorypressure] Generate Critical crash reports every 30 mins
We generate crash reports when the memory pressure level changes to
Critical, only if we've seen a Normal transition since the last report
was generated. This can cause us to not generate any reports for long
periods of time if the memory pressure level never returns to Normal.

Change the logic to also generate crash reports if it has been 30 mins
since the last report, even if we did not go through the Normal level.

Test: memory_monitor_unittest

Original-Bug: 57821
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414864
Original-Revision: 19f075d5f24112f68ced7cbac487eb3488acfaf6
GitOrigin-RevId: aa48d11b78f73c0efb0ed5cdba9e2bc1778c7050
Change-Id: Ib771b546490674036937dd6b0e95273f7095e566