[roll] Roll fuchsia [images] support board_tools
Adds support for a |board_tools| arg which can provide a list of tools
to include in the build output.

This is primarily meant for board-specific tooling that may need to be
distributed along with the build files in order to properly use them,
for example a tool to talk to the board in SoC recovery mode for
zero-state flashing or recovery from bricking.

Original-Bug: 56146
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412700
Original-Revision: 1f573da3e488dbcb96794a8bd8cd4647febeacf2
GitOrigin-RevId: 1e057cafb15d069ec3d10e7c01bbe6cb8d8faf1e
Change-Id: I97d82b0a3cba897e678796ac2b6f91aa45b6445d