[roll] Roll fuchsia [GN SDK] fserve.sh handles configured device info better.
This change removes the confusion around which device to use with
fserve.

Now, the command line arguments take priority.
if --device-name and --device-ip are not present on the command line,
then the configured device-ip is used if set.
if that is not set, the device-name configured is used.
if that is not set, it uses a device found with device-finder.

Original-Bug: b/163305432
Original-Bug: b/163304430

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415414
Original-Revision: 46f5d18b5405a1e0db749294fdc83b490be5790a
GitOrigin-RevId: 7d7a04a74e640c50dca7ec1810c24eeabfe8e801
Change-Id: I4ada61454ffd6c9b5055a48007191b0a42e3724e