[roll] Roll fuchsia [bt][fidl] Implement AudioDirectionExt server
Implement AudioDirectionExt server that sends ACL priority command when
SetPriority() is called. This is a short-term hack, and the command will
be properly managed in l2cap in the future.

It is currently possible for profiles to interfere with each other if
they request conflicting priorities. This will be fixed with a more
robust conflict checking system in the future, but it shouldn't be an
issue for now as we only expect to create 1 A2DP channel with a peer at
a time.

Reverse the ACL priority command when the AudioDirectionExt channel is
closed.

Create hci::BcmAclPriorityCommandParams as a temporary definition of the
ACL priority command until proper DDK support is implemented.

Add handler for vendor commands to FakeController. Tests can register a
callback to be called with any vendor commands received.

Original-Bug: 57163
Original-Bug: 57727
Test: in bt-host-unittests
FIDL_ProfileServerTest_ConnectedPeer/PriorityTest.ConnectAndSetPriority/*
FIDL_ProfileServerTest_ConnectedPeer.InboundConnectAndSetPriority

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414006
Original-Revision: 4f1eeafcd2a0ee96d1808d0f477186df707659dc
GitOrigin-RevId: 660cda2be8ffe84c46dcb1ee10746bd3a7c46baf
Change-Id: I9c8895a5bd26bf8140d0bb02bf6663865cbe49f7