[roll] Roll fuchsia [FDF] Exit gracefully when stop event completes
In component lifecycle stop event handler, driver manager suspends all drivers and
If the termination state is Reboot/shutdown, it also issues reboot/shutdown syscall.
At the end of the handler, closing the lifecycle handle causes the component manager
to kill driver manager. When driver manager which is marked critical is killed, the
root job watcher in kernel issues a reboot. If mexec is switched to use orderly shutdown,
this results in an undesired behavior where we want to mexec the system, not reboot.

This changeset exits the driver manager gracefully instead of closing
the lifecycle event handle. If we do so, kernel will not reboot the
system.

Original-Bug: 57448, 52627
Test: fx reboot. fx netboot after moving mexec lib to orderly shutdown.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415478
Original-Revision: 0885087a251702c73fe1b3d97841b47547ae0bbe
GitOrigin-RevId: 9f509aa9a4016909101009a9c9005477f3e83031
Change-Id: I15bb323a4eb87275873c4ff0a3cbcfddca4b3dc1