[roll] Roll fuchsia [kernel][sched] Fix a bug in deadline eligibility re-evaluation.
Fix an edge case in deadline eligibility re-evaluation and ensure
that the preemption timer is always set unless switching to the
idle thread.

When setting the preemption timer it is necessary to consider
deadline tasks that will become eligible before the current time
slice expires. The logic to do this had a bug that clobbered the
original preemption time and failed to restore it when a the
preemption timer was set earlier than the original preemption
time and re-evaluation didn't result in a thread change.

Update the logic to be clearer, restore the original preemption
time, and avoid unnecessary early preemption when newly eligible
deadline tasks don't have an earlier deadline than the current
task.

Finally, add a debug assert to check that the preemption timer is
always armed when selecting a thread other than the idle thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414866
Original-Revision: 9fdbf8eae6d817ccadae52c5a030bbc588b74d7b
GitOrigin-RevId: 18ef377eb9ca6adeeb4be5a39ad28f0bf9a3b25d
Change-Id: I002efde233089fa1b076b54ea0b332a47e22ef56