[roll] Roll fuchsia [fidlc] Allow flexible modifier for enum, bits
Remove the `flexible_bits_and_enums` experiment, mainligning the support
of `flexible` and `strict` modifiers for enum and bits declarations.

Adding a flexible enum, and a flexible bits example in the fidlc
testdata paving the way to proper goldens as we support this feature
throughout various fidlgen.

Expose `maybe_unknown_value` in the JSON IR which specifies the chosen
unknown value, should bindings whish to offer a simple way to manipulate
a generic unknown (for instance to send an unknown value to a peer). See
Id15bc3968410730180582672c6d50ac5c51aa6da which introduced this value
in the internal representation of the compiler.

Lastly, not updating the language grammar as it already has both
modifiers allowed.

Test: modified fidlc tests, goldens
Original-Bug: 7847
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415175
Original-Revision: 10c32827315618d3a114f9968bebdea10ee1e44e
GitOrigin-RevId: c50a47a250c310a1306478cdfa35c110403a8888
Change-Id: I74d86eedbbb66b8db4c1cd4ea87113a49c7e62eb