[roll] Roll fuchsia [fidl][fidlc] Clean up C union handling & more.
This cleans up the C generator's handling of unions and the the
allow-list for code that uses the C bindings beyond what is supported by
the generated code.

The developer visible changes are:
 - structs must now be marked [ForDeprecatedCBindings] for them to be
   part of the C bindings.
 - if protocols are not [ForDeprecatedCBindings] then Name and Ordinal
   constants won't be generated.
Previous CLs have cleaned up FIDL uses to take care of these.

Internally, the allow-lists have been changed so that legacy protocols,
structs and methods can be marked to be treated as if they are
[ForDeprecatedCBindings] even if they don't follow the all the rules.
Unions in structs and arguments are now fidl_xunion_t rather than an
incorrect generated type.

Encode and decode functions aren't generated if they would have to
encode or decode unions or tables rather than having them fail at
runtime.

This tighter, more comprehensive allow-list infrastructure highlights (I
think) all of the places that are pushing the boundaries of what we
support in the simple C bindings and hence give us a list to work
through. This should also insulate FIDL library maintainers who are
update libraries that are used from the C bindings from the binding's
constraints, so long as the types being changed aren't used from the C
bindings.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414138
Original-Revision: 0170c2ef7010753cda42531c1693046283e255f3
GitOrigin-RevId: b1925a1bee5a4601b506baea72fefa63e8dde3fc
Change-Id: I5e535736fc11307912276d365a1cc2e281780edd