[roll] Roll fuchsia [kernel][clocks] Skip trivial rate adjustment updates.
If a rate adjustment update is requested without a position update,
and the update request is "trivial" (e.g. it is attempting to set the
rate adjustment to what it already is), then skip updating the
transformation equations for a clock.

Original-Fixed: 57593
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413390
Original-Revision: 5bffc039dcd437a4806bc8c80ba665d7678ed397
GitOrigin-RevId: 6f04c0f90e6de12cc994fdcbf6a05ffaf0889403
Change-Id: I8846616771e809d2d285c8a706fa7833febec7cc