[roll] Roll fuchsia [audio][astro] Move enable gpio to controller
Having the control of the enable gpio in the controller facilitates
configurations where more than one codec is connected to the enable
gpio, as in the case in sherlock. With this we consider the enable
gpio an input to the codec and not in control of the codec.

Note that the board driver now hardware resets the codec via the GPIO
on initialization.

Test: tas27xx-test
Test: astro-tdm-test
Test: dvice-enumeration-test on Astro
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413878
Original-Revision: bb35f809c74187b5f66d2c3b0e4c7324c0860341
GitOrigin-RevId: 69e4325a9bdb0094e5db6a80340372407f514002
Change-Id: Ia60424d227fc55df55b0a503002d210cb8e1a778