[roll] Roll fuchsia [components] Monitor for process exit
Watch the process for exit instead of the channel. The runner
should not watch for the Lifecycle channel to close since this
could cause the runner to kill the component's Job incorrectly.
In particular this prevents us from killing a process which could
be marked critical to the root Job.

Original-Fixed: 56974

TEST=component-manager-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/411877
Original-Revision: 358e2f304771e5ca0c246ef6b5ba9dd45b734c06
GitOrigin-RevId: fc34b6fd3912905137911136fb1168620ba2e203
Change-Id: I337bb0f4d2ac054c3c111ce76264206abb4d74c2