[roll] Roll fuchsia [sshutil] Don't send a keepalive ping on startup
Previously Conn.keepalive() sent a ping as soon as it started, because
the logic to wait for a tick was at the bottom of the for loop, after
the ping was sent. By moving it to the top, we ensure that keepalive()
will only send a ping when the timer fires, which makes testing simpler
and easier to reason about, as well as more explicit.

It's probably not necessary to send an initial ping anyway, since
keepalive() is called immediately after the connection is established.
So there's not much likelihood of losing the connection within the
several milliseconds before that initial keepalive ping is received.
It should be fine to wait for the regular keepalive interval to elapse
before sending the first ping.

I doubt this will fix the flakiness exhibited in http://fxb/50198, but
it will hopefully make it easier to figure out what's causing the
flakiness.

Original-Bug: 50198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415919
Original-Revision: 7ece084be46236cce7e6ed703e950f6d39816c7c
GitOrigin-RevId: e1e600dfbc2fb76d4dd430e13dd22f719edfed9a
Change-Id: I9bbd89f74b47280315ddccf05f490466d3267c07