[roll] Roll fuchsia [testparser] Serialize empty test case list to empty array
Previously, empty "cases" lists in summary.json were sometimes `null`
and sometimes `[]` depending on the implementation of the corresponding
test parser. For example:
https://logs.chromium.org/logs/fuchsia/buildbucket/cr-buildbucket.appspot.com/8872159991166498784/+/steps/launch_collect/0/steps/0/0/steps/passed_tasks/0/steps/affected:QEMU__attempt_0_/0/logs/summary.json/0

It's very easy to mix up nil slices and empty slices in Go
and it generally doesn't matter which one you use, so instead of
changing the implementation of every language's test parser we can just
enforce in `Parse()` that we always return an empty slice instead of a
nil slice when no test cases were parsed.

The same thing is also necessary in tefmocheck, which is the other place
TestCaseResults are instantiated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416334
Original-Revision: 0e35833796c34da227e1cd551c8d389c514d9a3a
GitOrigin-RevId: e845e87884104649a286176728b805e74132407c
Change-Id: I4817026d4f44c6792f4e6d8b569a323eacb3e19a