[roll] Roll fuchsia [debugger] Clean up Register creation.
There was a lot of duplicated helper code to create Register structures
from data.

This creates a set of constructors for the Register class to initialize
from different types of sources, and the helpers are removed.

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416157
Original-Revision: ece0be7251fbe45d5317b41270580de10efcef75
GitOrigin-RevId: f4546993d9d68de8dce79152c2b6b3db55981d71
Change-Id: Id8f9e5591abfa661aae650dfee3c2e36f0288b3c