[roll] Roll fuchsia [swd-e2e-tests] Wait for a reboot if `update` fails
The `update` command has another reboot race, where it's possible that
after an OTA was successfully installed, the reboot process could tear down
the services that `update` was talking to before it receives notice that
the OTA successfully happened.

This patch changes the `Updater` service that upon error, wait 30 seconds
to see if the device reboots anyway. If it does, treat this as if the OTA
sucessfully installed and the device rebooted as normal. Otherwise, err
out.

Test: manually changed `update` command to always error, and observed the
test still succeeded. Replaced `update` with `exit 1`, to make sure it
errored out.

Original-Fixed: 58086
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416358
Original-Revision: e349adc8123d5e98d3fc468790560618ed53b013
GitOrigin-RevId: be59a81728d7bfa16d4f305374bc38424f1a0d74
Change-Id: I3a6cb9547f0e42202a16b71056bf38d27dd06279