[roll] Roll fuchsia [timekeeper] Move timekeeper to ergonomic inspect.
This keeps approximately the same scope of inspect data as previously
but starts to add structure to both the code and the output format,
produces more inspect data as events occur rather than lazily, and
drops some of the less relevant fields from clock details.

Note that timekeeper doesn't really match the design asssumptions of
the fuchsia-inspect-derive library and so in our case the ergonomic
inspect isn't very ergonomic - things are still quite verbose and we
have allocated additional memory to store inspect values in a place
the component could, but does not, read from memory.

Improvements to fuchsia-inspect-derive (e.g. fxb/57995) may help, but
as things evolve its also possible we'll want to stop using ergonomic
inspect and go back to the raw API. Even if that were the case, the
high level structure of the code and of the inspect output should not
need to change substantially.

Original-Bug: fxb/56872

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414183
Original-Revision: d1b9ee0206ee420a49f47f6edfb75d0c576fcb54
GitOrigin-RevId: 1cce1f948bf01e528564c083d257ba158f3d563f
Change-Id: I21f52a2d38b2edefd383e5c9a7375803a3761d5e