[roll] Roll fuchsia [audio][astro] Fix channels configuration
We allow masking of channels as they are provided to the
driver in the ring buffer. The enablement in the AMLogic
SoC was incorrectly using the second channel (Left) which
would be filtered out if the masks for channels to enable
only had setup the first (Right) channel. This worked when
no filtering was applied in the driver (as in tones via
audio-driver-ctl which does it in software), but failed
when filtering is applied in the driver (as in
signal_generator and audio_core). The fix is to configure
the correct (right) channel as the only mono channel used
and make sure the codec also uses the right channel (asserts
added).

Original-Fixed: 58173

Test: tas27xx-test
Test: Manual check of sounds from button pressing
Test: signal_generator --sine=440 --amp=0.95 --dur=1
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416322
Original-Revision: 70019c52696bae5b093088146038faa59417af13
GitOrigin-RevId: 69caac45a7f2c310c2a436ea126ec6e7e7b9f726
Change-Id: I8ffa68207080bb59280d96c4456061eb8771468b