[roll] Roll fuchsia [fatfs] Fix shut down race.
When you shut down the execution scope, it can return before all futures
have actually been dropped, but we were relying on that to properly
clean up.  So, to address that, we stop shut down from relying on
everything to have been dropped and just flush instead.

I also took this opportunity to remove one of the levels of indirection
for the filesystem.  It was only necessary because the get_filesystem()
method returned Arc, but it doesn't need to.

Original-Bug: 53796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416540
Original-Revision: 9420e2e7395fcf5848f39cb9817ef7641dd62abf
GitOrigin-RevId: b271508a975c50e092dc630d471192a37550c887
Change-Id: I1b3b8f2988060c3d64b17e8eac0ee2195d857353