[roll] Roll fuchsia [scenic] Refactor swapchain format selection.
This moves pipeline warmup to the time when swapchains are created
in order to not hard-code the image format.

Swapchain format is now determined by a single list of preferred
formats in the BufferPool class. The preferred format is still
B8G8R8A8 as before but adjusting it is as simple as changing the
order of formats in this list.

This can be extended to support a board specific preferred format
in the future if needed.

Original-Bug: 42571
Test: fx test gfx_pixeltests gfx_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416197
Original-Revision: df0bf2bd4ae26d6df52fecadea69ae5a0755e6aa
GitOrigin-RevId: 65246c53c940cff1c5c28e81da9c8855632b73ee
Change-Id: I2f250bad2496519e674c9bffd707a37aaaf222b5