[roll] Roll fuchsia [omaha-client] Extract check-now functionality
This pulls out the fuchsia.update.Manager.CheckNow()'s implementation
into a sub-function. This allows us to add metrics tracking the result
of the CheckNow() method in a future CL.

While this is mostly backwards compatible, it makes two small changes to
some error edge cases, which caused the Manager channel to be closed
without a response being sent:

* it's possible that the monitor type is not a proper channel, which
could result in the `.into_proxy()` method failing. This was changed to
returning an `InvalidOptions` error to the caller.
* it's possible we could get an error adding a monitor to the
monitor_queue if it was dropped. This was converted into an `Internal`
error.

Test: I'm not actually sure how to actually produce these errors, so I
think these cases are largely theoretical. Looking at the rest of the
Fuchsia codebase, most other users either close the channel without an
error (with `.into_inner()?`), or just panic (with
`.into_inner().unwrap()`). Since it's pretty fundamental to fuchsia
working, it seems best to communicate an error occurred to caller, and
also not panic the omaha-client process.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416194
Original-Revision: 3ba699e46729bc61aa2a5503653780101b10e048
GitOrigin-RevId: 591d5a9d4f05088177ada8286755b8f09750ab58
Change-Id: I04d7ee039171825d5aec2beb02434edb7a7f8744