[roll] Roll fuchsia [mdns] ensure query responses are sent immediately
This CL fixes a problem in which responses to PTR queries were sent a
second late, because outbound resource records weren't getting
promptly flushed out as messages.

'Agents' in the mdns service can 'send' resource records and
questions, but those items are accumulated in the Mdns class until
Mdns decides to 'SendMessages', which creates messages from the
accumulated items and transmits the messages. That happens at various
points that make sense for normal agent behavior. The problem arises
when an agent sends items asynchronously with respect to the times
when SendMessages is normally called. InstanceResponder, the agent in
charge of publishing service instances, does this when the client
responds to a GetPublication call.

This CL adds a protected method (FlushSentItems) that causes Mdns
to SendMessages. InstanceResponder is currently the only agent that
calls that method.

Test: fx test mdns_test
      Added a test that verifies the combined behavior of Mdns and
      InstanceResponder.

Original-Fixed: 58141

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416838
Original-Revision: cd38efaaf4827500b690639b2cffa1ae090cb791
GitOrigin-RevId: 0929bf9d3f0981a28effebb1047523727686efc4
Change-Id: I78da961329223bb1021870e720e57945ac8a8775