[roll] Roll fuchsia [ffx] add sdk target and enable cross-builds
In order to build for any SDK target for both x64 and arm64 we need
to enable cross-builds by way of using is_host instead of
current_toolchain comparison. This is still a little imprecise, as
really we only build for linux and mac targets today, but it's
approximately intent oriented instead, and simpler to spell.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415482
Original-Revision: bc1fd913d3859d7c84200a1737b92364d1c90175
GitOrigin-RevId: 14c6cdb0039faf247b7b04d73b456fe57fc45475
Change-Id: Ic7442b4d62ee5347e6060dc9b6859d53d85b793d