[roll] Roll fuchsia [power_manager] Add ThermalShutdown node type
Creates a new node called ThermalShutdown. This node type will handle
the simple task of polling a given temperature sensor and checking the
temperature value against a configurable threshold. If the temperature
exceeds the threshold, the node will enforce a system reboot.

This functionality currently exists as part of the ThermalPolicy node,
but there are a couple reasons we need this in a standalone node:
  1) On some products we'd like to use a basic thermal policy like this
  without all of the extra functionality (and dependencies) that come
  along with the ThermalPolicy node. Currently we're forced to enable
  the entire ThermalPolicy node and use hacks to disable the
  functionality of that node that we don't want.

  2) We need the ability to monitor multiple temperature sensors for
  critical shutdown temperatures. The current ThermalPolicy isn't
  designed for this. By having the functionality brought into a
  standalone node, we can enable an instance of this new node for each
  temperature sensor we'd like to monitor.

Test: fx test power_manager_tests
Original-Bug: 52898
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412626
Original-Revision: d9ed44c1f56731346c58f1f368ebe09410270772
GitOrigin-RevId: 56b1ddca2b3d761b63d0644a4a8f9583dcdbe575
Change-Id: Ia614c0ba5686ce053d8e62a3e22c09a2dffb2c48