[roll] Roll fuchsia [build] Support for emitting GSYM files
This is not quite ready yet because llvm-gsymutil currently doesn't
correctly handle symbols with zero size, but it allows us to start
experimenting.

For now, we only support emitting GSYM for target-side ELF binaries.
We are going to expand this to host-side ELF and Mach-O binaries once
the tooling is stable.

Original-Bug: 58294
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416214
Original-Revision: 24e2367f28494879636f5102aa0050bd5d614991
GitOrigin-RevId: 56aea5b3b27fe1cd81131a3a0dbe08f6d51d3f7a
Change-Id: Ie8f5ecf2c7c247a23866de9880a5f2b8b6273534