[roll] Roll fuchsia [power_manager] Use dynamic strings for node name
The Node trait defines a function `name()` that is intended to return
the human-readable name of the Node implementation, mainly for logging.
The function is currently defined as returning a `&'static str` but
this severely limits the flexibility that is sometimes required. For
example, the TemperatureHandler node returns "TemperatureHandler" as its
name, but it would be better described by the temperature driver that
it's bound to as well. With this change, the string returned by the
`name()` trait method is a dynamic String that allows the Node
implementation to use a more detailed name that wouldn't be possible
using a static string.

Test: fx test power_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/413397
Original-Revision: a90d1feeb3fc4a9d6db4fefe9d2f1cbd3a94343f
GitOrigin-RevId: aaad3995d2d090b9d8da044b470552c888711fa5
Change-Id: I731030e6fbca31db8b9e6faca3b85adf9bef4989