[roll] Roll fuchsia Reland "[Settings] Use ExternalError when handler calls fail"
Original-Fixed: 58214
Test: fx test setui_service_tests setui_client_interface_test

This reverts commit 671b19bdded6fa1b20302bc2543cf56f01f81586.

Reason for revert: Relanding after fixes.

There was a line that changed self.stored_stream that was moved up above
the set_volume and set_mute calls in set_volume in order to ensure it
still got set even if the call failed. It was causing the
stored volume level to be the same as the new volume level, and it
was not detecting the volume as changed after the first change. This
fixes it so that the stored stream is saved after comparison but before
the errors.

Also adds a test to make sure consecutive volume changes are registered.

Original change's description:
> Revert "[Settings] Use ExternalError when handler calls fail"
>
> This reverts commit 73518d0eadd4c90cda67fba76eeea9bb47e0c06e.
>
> Reason for revert: Breaks volume control
>
> Original change's description:
> > [Settings] Use ExternalError when handler calls fail
> >
> > This ensures that ControllerErrors are returned in the controllers' handle trait methods.
> >
> > Fixed: 57172
> > Test: fx test setui_service_tests
> >
> > Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414717
>
> TBR=wxyz@google.com,cfukuhara@google.com,paulfaria@google.com,setui-review@google.com
>
> No-Presubmit: true
> No-Tree-Checks: true
> No-Try: true
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416394

TBR=brycelee@google.com,wxyz@google.com,cfukuhara@google.com,paulfaria@google.com,setui-review@google.com

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416615
Original-Revision: ea943682637ee31a5ca3f3200ead792337136c3b
GitOrigin-RevId: 6ff9ea9b1c3dd5de063d7d71eb126497e19e5b22
Change-Id: I35b684c68abc212fe0e807e82daaf4fd971c1fd4