[roll] Roll fuchsia [audio][sherlock] Fix rates issues
The unified AMLogic audio driver was setting the audio clocks
with the same logic as in Astro, this means dividing by 2 the
mClock division for 96kHz configurations. In Sherlock before
unification the clocks were configured such that instead of
mClock, sClock would be divided by 2 for 96kHz configurations.
This CL changes the clock parameters in Sherlock to match Astro
fixing the bad division in the unified driver. This CL also adds
an assert to make sure the divider parameter is as expected.
This change has the nice side effect of lowering the HIFI PLL
rate and making both products use the same audio clocking scheme.

This CL also fixes rate setting of the codecs (missing loop).

Test: aml-g12-tdm-test
Test: Youtube video manual playback.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416955
Original-Revision: cdd476cefcdc8b3ffbfe06220653e2b0f2ad751b
GitOrigin-RevId: 7d5bca735ce6faae5458ce2ae0002c235f3c1f4c
Change-Id: Ib891395241bb297ee32b7dc6ef9381f80d97446b