[roll] Roll fuchsia [wlan][brcmfmac] Only get wsec_key iovar for WEP tests in auth_test
The wsec_key iovar only succeeds when the security mode is WEP. It
does not make sense to check this iovar for other security modes,
such as WPA1 and WPA2 since the handshake to establish the key is not
simulated yet.

This change also removes the use of getting the 'wsec_key' iovar using
IovarsGet() which is the final use of that function outside of
sim_fw.cc and iovar_test.cc which are the only places where it should
be used.

Test: fx test auth_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416707
Original-Revision: 27cd816b52c217d80a64efd810c7fb4cda360c04
GitOrigin-RevId: 283205d8cf5c8b0c9621eb76e06380541bf261af
Change-Id: Idbebfb187b2c2bc4bc9553ac99baed84fdcbcd7a