[roll] Roll fuchsia [wlan][brcmfmac] BCDC ioctl buffer handling improvements
= Warn when transmitted message buffers are too large

BCDC ioctl is bounded by BRCMF_TX_IOCTL_MAX_MSG_SIZE for txctl and
BRCMF_DCMD_MAXLEN for rxctl. The former is never more than the
latter. The transmit limit is enforced by truncation and a warning,
but trasmission of the truncated message will continue. This is
because in some cases the length of the buffer may be intentionally
longer than the transmission limit so it may accept a long reply.

= Use rxlen_out to determine actual number of received bytes

While the firmware *should* always return the same number of bytes in
bcdc->msg.len, it's possible that it doesn't at times. This type
of abnormality is reported by the rxlen_out argument of
brcmf_bus_rxctl.

= New test

Added a new test to ensure BCDC fails fast when the caller provides
a buffer that is too large.

Test: fx test brcmfmac-sim-test and manual scan, connect, ping, disconnect; and ACTS tests WlanPhyComplianceABGTest, WlanScanTest, ConnectionStressTest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404141
Original-Revision: 124cbc3e8ee7f7b70b26312b5805da47e98e73e4
GitOrigin-RevId: 8929d0d2f8e486ba030c6015c26a1513ad23d78b
Change-Id: I8f553e28069f390eb2c77cc2aa49dea4aab47868