[roll] Roll fuchsia [brcmfmac] Wrap wlanif histograms in feature flag.
This change also reduces the verbosity of the logging when an attempt to
retrieve the detailed histogram fails, as requested in fxb/58065.

Test: a new test that disables the detailed histogram functionality
passes, showing that the stats query will continue to work when the
feature is disabled; existing brcmfmac-sim-test tests that exercise the
detailed histogram functionality continue to pass.

Original-Bug: 58065
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416680
Original-Revision: 237f2622a6803368a55176622346ba8006938537
GitOrigin-RevId: f63fa8732b74fd9d875d510be642cce1957cd9ed
Change-Id: Iebf3487cc9364fb4c5739d62891d03111966d348