[roll] Roll third_party/openweave-core [openweave] Fix rendezvous over ipv6
The ingress and egress interfaces for link local ipv6
packets are always set to 0. This is because we dont support
IPv6_PKTINFO and it causes the outgoing packets to get lost.
Use the scope_id from the PeerSockAddr struct so that packets
are sent to the same interface as incoming packets.

Original-Bug: 58151
Test: Verified that rendezvous from weave-device-mgr works.

Original-Revision: 954301efab6c031518130784d8a0e3901ce7c3fc
GitOrigin-RevId: f50536b0bec1a8b42566cdb95327aea65aa93987
Change-Id: I2027bd6a3991e87b9ab6cf17e3f045627a68bb92