[roll] Roll fuchsia [ot-stack][test] Removing flaky part of a test
Removing unnecessary part from TEST(Alarm, RemainingTimeBothAlarms).
The test intent is to set both micro-sec and milli-sec alarms
and to sleep for some time, and check remaining time on micro-sec
alarm to make sure it's correctly reported.

Since in infra, the test may sleep longer than programmed, it may result
in more possibilities to happen rather than intended case. The test detected
some of those scenarios and then checked the results against it.
However, there are many different such scenarios for this test and taking
care of all those adds to unnecessary complications and resulting in flakes as well.

The test intent is still met. The commented out part of the test - that
is making sure that remaining time is reported as 0 when alarm as fired
is already tested by other tests (clearmicrosectest and clearmillisectest),
so we're not loosing coverage

MULTIPLY: threadstack-unittests: 200
Test: fx test threadstack-unittests
Original-Bug: 58032

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416795
Original-Revision: ba10d327fb676f7d58e397386132e6b485a7b4b4
GitOrigin-RevId: ffb159fb7ee2e74e8736fc541945322e88d662e3
Change-Id: I204656123e693a2d76d64ea1951cc273c4aa557c