[roll] Roll fuchsia [aml][pwm] Add ability to protect pwm channels
Some pwm channels may be in use by higher ELs or be running
in a state configured by bootloader which is required to maintain
safe operating conditions.  This CL adds ability to mark
channels as protected in the board file.  This flag is checked
by the pwm driver (and protocol) to block initialization of and
access to channels marked as protected.

TEST: Tested on astro while monitoring PP800_EE rail
Original-Fixed: 58137

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416819
Original-Revision: d2d7dd270b70ac4bf3c061806471fcd74ba6815c
GitOrigin-RevId: f5d9e59d83ef133a53bcd7e6f3f6cd38cb629d05
Change-Id: I6a3fae095d892b23a960bc6160dd47299d5051d7