[roll] Roll fuchsia [kernel][vm] Remove non-const VmPageList methods
These non-const methods allow for modifying the markers in place, which
is generally an unsafe operation as it does not allow the page list to
cleanup empty nodes or perform other modifications.

In the occasional scenario where in place modification is desirable the
RemovePages method exists and is the intended method to be used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416535
Original-Revision: bd31cf5772ad75981031f5e92a5a36a288257ef9
GitOrigin-RevId: f7599bf9e470f54f60cdc1458622ddf871a7fd77
Change-Id: I85920e1fe97eec71afa672c283cdcbbdebb7222d