[roll] Roll fuchsia [sshutil] Have testserver.stop() wait for goroutines to exit
We've been seeing flakes in the
"TestKeepalive.disconnects_conn_if_keepalive_fails" test case.
That test case starts an SSH server and creates a connection to it, then
then shuts down the server before sending a keepalive ping through the
connection. The point of shutting down the server is to ensure that the
keepalive ping fails, which should in turn trigger a connection closure
in the keepalive() method. The flakes happen when the connection doesn't
get closed as expected.

By adding logging and running the test many times locally, I determined
that the assertion failures happen when the server actualy *does*
respond to the keepalive ping, so the keepalive logic doesn't close the
connection. I believe this is possible because the server listens for
new connections in a goroutine, which in turn launches child goroutines
to server any new connections, but stop() doesn't wait for those
goroutines to exit before exiting itself. This means that even after
stop() is called, new and existing connections could still be served.

I addressed this by adding a sync.WaitGroup field to the server that
tracks both the listener goroutine and all the child connection
goroutines. Now stop() will block until all of those goroutines have
exited and it's no longer possibly for any new requests to receive
successful responses.

I ran the test with the fix 5000 times locally on my Mac and 2000 times
on my Linux workstation, and didn't get any failures. Previously, I was
able to get a failure pretty consistently (especially on Mac) by running
the test 1000 times.

Multiply: sshutil_tests
Original-Bug: 50198
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393475
Original-Revision: 2cbd6f0c0e8c7d3d270120a5224282543cb7b835
GitOrigin-RevId: 4d6b3100b02f1e44fd83141c817d28912c7898d6
Change-Id: I7c73faa3ce9ef848b3178698e2e1a4ae596e0188