[roll] Roll fuchsia [storage] Issue 57889: Support flush for the rust VFS
This change is adding flush support to the Rust implementation of a
block client - RemoteBlockClient. It comes with additional tests that
ensure that flush is being registered by the server's block stats and
that a fake server properly receives the flush requests.

Test: To test run the block-client-tests as below
  - fx set core.astro --release --with-base='//garnet/tests/benchmarks:garnet_benchmarks'
  - fx test block-client-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415378
Original-Revision: 15076c86d38031131f931c039fe94dd3ff209af3
GitOrigin-RevId: 445e87887ac0195a07a16aa558b9ae95ed6763e1
Change-Id: Ifef3bebd0725ae2d4921e542ec0fb99cef7a1603