[roll] Roll fuchsia [tap] Add doc.go and fix typos
When I first encountered this package I was very confused by the name. I
assumed it referred to TAP, the Google internal testing system, but it
actually refers to the Test Anything Protocol, which is open and not
affiliated with Google. So I added a doc.go to clarify.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417780
Original-Revision: 72f5bbd563ae96eb0af1c6d358fa447aeb7e6fea
GitOrigin-RevId: b1e8b0f64dcea37632baefb16427a8b8b6a9877d
Change-Id: I07949ecf8e95a276682294a3fa1f4f274f5a110f