[roll] Roll fuchsia [flake][bootserver] Move "netsvc: start" later
Tests rely on the output of "netsvc: start" to determine when to start
communicating with the target. Because the advertisement happens later
than the print, there's a race with a host-side running of (e.g.) netls.

Move the print later so that it's once the advertisement has happened
once.

Multiply: host_x64/bootserver_test_pave_no_bind_failure
Original-Fixed: 57252
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417322
Original-Revision: 1bc12fa71053347c91d07352881937a863fec637
GitOrigin-RevId: b9451bf3f5bd0e229c578ac67b90bbf4ec9ebb51
Change-Id: I907a2c78b9d50f99d22972a2a87e3cd32798725f