[roll] Roll fuchsia [zx][public] Add conversions between Zircon time and C11/C++17 timespec
This is a slight departure for the lib/zx API, which mostly only
uses "language" features and not "library" features.  This
introduces a header-only dependency on <time.h> (C) and/or
<ctime> (C++) to declare the `struct timespec` (`std::timespec`)
type.  lib/zx headers already have a header-only dependency on
standard C++ library headers for <limits> (std::numeric_limits),
so this is not a practical change in the dependencies.

In <lib/zx/time.h>, the new functions using `std::timespec` are
gated on compilation being in C++ 17 mode.  (`std::timespec` was
added to the standard in C++ 17.)

In <zircon/time.h>, the new functions using `struct timespec` are
gated on `#if __has_include(<time.h>)` so they are not available
in environments (e.g. efi) that don't provide that part of the
standard C library API.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417330
Original-Revision: 0b76528f14bdefecc7bba39e7b224367798500b0
GitOrigin-RevId: c6f3b5f717ada54a01740181d9a7609c9bd1b9d1
Change-Id: I60d1a08074f3f2a6031ddeaf38dbaa160145dcdd