[roll] Roll infra/recipes [sdk] Move gerrit link to the step level
During execution, the gerrit link does not appear at the parent step
until the parent step is finished executing.

Move the gerrit link to the "create CL" step, so the CL can be opened as
soon as it is created.

If the build fails, the gerrit link will still be surfaced in the
summary markdown.

Original-Bug: 58445
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/417914
Original-Revision: 81dfe19d655404750ddc16be30f1b2db09ca27ca
GitOrigin-RevId: 37465d2f513893fa9f35fc38fca110f2b051709b
Change-Id: I67c977a38a9205804ab01a3853d0083932509423