[roll] Roll fuchsia [dhcpd] Continue on unrecognized option code
When parsing a byte buffer into a DhcpOption::ParameterRequestList, an
unrecognized DHCP Option Code should not cause an error and parsing
should continue with the next Option Code. This change addresses a bug
in the DHCP server where an unrecognized Option Code in the
ParameterRequestList would lead to the server failing to return values
for any of the requested parameters.

Test: fx test dhcp-tests
Original-Fixed: 54345
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417887
Original-Revision: 1deadf19f39a68987ec0ba95c001bef448033283
GitOrigin-RevId: 5a8ab380bde5bb6b5417cd3ec6e96f8c57919be4
Change-Id: I2a5915a6fd55a76494ca9c5c341953ca659c0bae