[roll] Roll fuchsia [bt][a2dp-sink] Add peer session setup
Move session setup out of ConnectedPeers so that the streams and tasks
related to the A2DP peer are started on first connect.  Move AVRCP task
to the peer session generation function.

This allows us to separate audio session creation from the AVRCP task,
and pass session ids to the stream media sink tasks, allowing us to tie
the streams that are generated to the media session for the peer.

Original-Fixed: 42976
Original-Bug: 43545

Test: updated tests in bt-a2dp-sink-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417894
Original-Revision: bd3655796d2539bb3a1146ad7206ecaa407e7e42
GitOrigin-RevId: e7de604e10e6f492bda2b2481569e542725990c1
Change-Id: I170fd605d91b2dbcccecc07d7463b3335ca45235