[roll] Roll fuchsia [brcmfmac] Set bounds on error injection buffer
When specifying a replacement buffer, instead of just storing a
raw pointer to the replacement value specify a pointer to a vector
so that we can place a proper bound on the values that are
accessed at runtime.

Also add a test that verifies proper behavior for various inject
buffer lengths (greater than, equal to, and less than the expected
length of the output structure).

Original-Bug: 51987
Test: Included (err_inject_test)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417886
Original-Revision: f96c03c89912849269d5daaf201a9eff45ebfedf
GitOrigin-RevId: a7257477ea8c72b8535be56ed349ab42e85a2485
Change-Id: I056ebb05754fa5d3e096b0f08c4fd9ce6a6f57e4