[roll] Roll fuchsia [driver_host][lsan] Fix leak
This leak was intentional, but cleaning it up doesn't seem like like it
would be a big deal, so I upadted it to do that. I used a static global
rather than returning a context/object as it was more conveniant.

Original-Fixed: 58496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418291
Original-Revision: ca9ab95767071701291b16f5bee3c3d1be1756a4
GitOrigin-RevId: 604380f37c508ad852df79788d39345eda2ca846
Change-Id: I433aef58ff2c1447dbd8d3afaf002927b751ecf5