[roll] Roll fuchsia [debugger] Update the module cache on init.
Updates the module cache as soon as the dl_debug_addr is known.
Previously the way we executed when launching a new process updated the
module list, but not when attaching to an existing process. This meant
the module cache was out-of-date.

Update the module cache when requested by the client. This shouldn't be
necessary but allows the user to "fix" future issues along these lines
where the list of modules is out-of-date, and ensures the frontend and
the backend are in-sync.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418314
Original-Revision: af6b374a1392289e98f466cf9a4a357f9098757a
GitOrigin-RevId: 11138c164083bcaf36fd21fcffcce5234b2e0258
Change-Id: I0ebe3f6d67b812d58135f12c622afb8ac3a911ac