[roll] Roll fuchsia [time] Preliminary time.external implementation for roughtime
Adds a basic time.external.PushSource implementation for roughtime
with only WatchSample implemented.
In a PushSource implementation, the implementation decides when it
should poll for a new time sample. For this implementation, the
time source only schedules a poll for a new sample if it has an active
WatchSample request. The advantage of this approach is that it does
not result in unused polls while the deprecatedtimezone protocol is
still in use.
The implementation currently assumes there is only a single client
and uses a very simple retry strategy in which time samples are polled
again a second after a failure, or 30 minutes after a successful poll.
Note that unlike the deprecatedtimezone implementation, this does not
update system time by itself - it only reports what it believes the
time is.
Up next: Implement WatchStatus, update the retry strategy to use the
exponential backoff with cap strategy currently in Timekeeper.

Original-Bug: 56833

Testing: fx test network-time-service-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/415832
Original-Revision: f43d162b3971efb5e0d7fc04518cc11be38f2cb4
GitOrigin-RevId: 70592c648de3aa7bf4d67817051438ed5013b046
Change-Id: I61025c1df723881337c57f96274c38216aa7528e