[roll] Roll fuchsia [kernel][uart][pl011] optimize the pl011 TX routines a bit
In the core loop to write a string to the uart, read the transmit
fifo empty flag bit to determine if we can safely stuff 16 bytes into
the data register at a time. Restructure the code to do this. This
reduces the number of times the VMM needs to trap and vm exit by reading
the flags register by approximately a factor of 16.

Drop the uart spinlock every iteration around the loop to try to
cap the max time we spend with it disabled. With the above optimization
it mitigates the acquire/release cycle to every 16 bytes.

Make sure we actually enable the fifo in the LCRH bit at startup. QEMU
doesn't seem to do much with it on the TX side, but it does seem to
actually attempt to better emulate a RX fifo by actually allowing the
generic qemu uart layer to shove up to 16 bytes into the dev's fifo and the
RXFF bits are appropriately set. No apparent change of behavior on our
part since we're still reading data out of the fifo until RXFE is set
and it's either 1 or 16 bytes based on if fifo is enabled. If we ever
see a real PL011 implementation however, the fifo will be a big win.

Add a few wrapper functions for common tx and rx mask/unmask routines.

Tested: boot and run on qemu with and without irq driven tx. Verify that
everything seems to work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417359
Original-Revision: 4f9231d0c7d00c00cdfcc3656dbf079efb1c5b9a
GitOrigin-RevId: 08aa44c05a4b81decd34dcc49c105f6f1c1a3034
Change-Id: I14b1a9d2c7cb7391fb301dfa3e4eaf373226f40c