[roll] Roll fuchsia [time] Remove redundant log
Timekeeper and network_time_service both output logs when an attempt to
retrieve time is made. This removes the log in network_time_service
which indicates an attempt has started, and reduces the severity of a
log that is output repeatedly in case internet is not available.

Original-Bug: 58390

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417997
Original-Revision: 5b924d6ca36d9b12b8f7de8eaf9ae9bf95df81d9
GitOrigin-RevId: 11b9e9467bdc64ec5c4cc0b6f84902bcaf7b763b
Change-Id: I2cac91a78545631c6a69f9773259c206f82fdea2