[roll] Roll fuchsia [storage][volume_image]: Add decompressor.
In order to check validity of the the sparse image header, the data
needs to be decompressed so the total length can be retrieved, in order
to prevent any sort of truncated file that might look fine, but extent
descriptors with out bound pointers.

Move LZ4result to its own header, so it can be shared between compressor
and decompressor.

TEST=storage-volume-image

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416401
Original-Revision: bf21b651a0460895efa96c257bf57b56c967e0fa
GitOrigin-RevId: e29983458a20a86865d11352a85c826600e7ce0e
Change-Id: I1113e71d17122b3bc388eabd7c826edf148552da