[roll] Roll fuchsia [time] WatchStatus implementation for roughtime time source
Adds a basic WatchStatus implementation for roughtime that piggybacks
on the update loop for WatchSample and declares whether or not it is
healthy based on whether the last time sample succeeded.
Also updates SampleWatcher as a template so that it can be reused for
watching Status as well.

Original-Bug: 56833

Testing: fx test network-time-service

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414383
Original-Revision: 7693682f83b0836faa00382d0b10e3493f35bbf7
GitOrigin-RevId: d92146adb322f183b00c809a821e2eaf86c66112
Change-Id: I40f1367c96444a86645d8f5dfa90525a4dd88309