[roll] Roll fuchsia [fs_test] Support blobfs
Adds blobfs as a supported variant for the fs_test framework.

Blobfs is intentionally not added to fs_test::AllTestFilesystems, since
it is not expected to run as part of the regular fs_test suite. Instead,
blobfs will run its own integration suite, using the fixtures provided
by the fs_test library. (Blobfs is not a general-purpose filesystem,
which is what the fs_test suite is designed for.)

Tested: Built and ran a test component in a child CL.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417738
Original-Revision: b902cff8c002b315a9020b76c4a378d4c4de57db
GitOrigin-RevId: a9bbd014cfbb24688bd3fe839318f20429e0fdc8
Change-Id: Ie5f6075d6164edec1d738ce2c6ded18e61c9c85f