[roll] Roll third_party/go [fidl][go] Isolate fidl host test impact to fidl package
This CL reverts a previous CL that added a dummy-with-panic
implementation of the syscalls needed for fidl host tests and instead
implements a redirection layer in the fidl package with system specific
implementations.

It is still necessary to make some changes to syscall/zx because handle
syscalls are accessed via methods on handles, which need type and method
body separated into different files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/417054
Original-Revision: a05726e70856de98b624ef914cfe95ecd8b7bb74
GitOrigin-RevId: aa47c9b4a24d78f10159f21bfb5b8b2f5daca046
Change-Id: I856fcaa58fde34c791b50102dfa4a25bfc02baf8