[roll] Roll fuchsia [msd-vsi-vip] Fix WaitUntilIdle().
It appears possible for IsIdle() to become false for a brief
period after returning true. This results in occasional test
flakes.

This change modifies WaitUntilIdle() to return after the first
true result of IsIdle(), which matches the implementation of
the reference drivers.

Ran the tests for half an hour and the flake appears to be fixed.

BUG=58070

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418536
Original-Revision: a6100616c095f21f0bb2b75d11196b9573f433dd
GitOrigin-RevId: aeb9e767e4c06bbd56bddea2814c956064f90652
Change-Id: I022bd2954c62fb9994a50ad2e3430d5e045a558e