[roll] Roll fuchsia [flatland] Add a queueing interface for UberStructSystem.
The UberStructSystem::UberStructQueue class is the intermediary
between Flatland instances and the UberStructSystem.

Flatlnd instances queue UberStructs for session updates through an
UberStructQueue object. The UberStructSystem pops these UberStructs
when told to perform session updates.

At the moment, this class wraps a mutex and queue, but in the future
it could be entirely replaced with a lock-free queue entirely. Issue
57745 tracks progress on this.

Because the UberStructQueue is managed by a shared_ptr, the
UberStructSystem can also detect Flatland instance destruction by the
reference count on the shared_ptr and de-allocate resources
appropriately, rather than depending on the Flatland instance for
this signal, which would introduce the need for a lock on the map
that holds the UberStructQueues.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414115
Original-Revision: 2f1847775dd07863c78aaee165d7f08ef99ea936
GitOrigin-RevId: 3cab21d1cdf7915d3907e838ddfcce894bee02f7
Change-Id: I6cc44b53ba56102fb94ffe48486a175ac42b859c