[roll] Roll fuchsia [bt][acl] Ignore out of bounds completed packets
When processing the number of completed packets, a larger number of
packets than was actually included in the packet could be specified,
causing data that was out of bounds to be accessed.

Calculate the number of packets' worth of data and do not process for
counts that are not provided.

Original-Fixed: 58592
Test: modified ACLDataChannelTest in bt-host-unittests to exercise

Multiply: bt-host-unitests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418774
Original-Revision: 19f2f23d396041d0efb6f2b2bd378a353152fea2
GitOrigin-RevId: 27ecdb7e14f1857917b855b2253bd5cd2aa14cbe
Change-Id: I81cda026c70fc5bac2a3870aa965371e9cbf3706