[roll] Roll fuchsia [audio_core] Use new overflow and underflow metrics
This CL refactors the Reporter to emit the new metrics defined in
fxr/416315. Currently I've implemented just the OutputDevice and
Renderer metrics. Future CLs will make use of the OutputPipeline and
Capture metrics.

Note that the OutputDevice session never ends. I expect this to change
in the future once ring out information is fully propagated down to the
output device and we can idle the mixer when there is nothing to mix
(this is fxb/50513).

Tested by manually checked the output of inspect metrics, which mirror
the information sent to cobalt. Also some unit tests.

Test: fx test audio_core_unittests
Original-Bug: 58121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416296
Original-Revision: 898b3948ecb2799c15325046150795721124be16
GitOrigin-RevId: 9e606f463b5257f6ad65479d1330436452586580
Change-Id: I95b10612a62467572bf691ef1f4dc3d3dae20358