[roll] Roll fuchsia [camera][imx355] Initial imx355 driver
This CL contains the initial version of the IMX355 driver.
When tested on a luis platform board with the enabling CLs to follow,
the IMX355 responded to I2C requests and, when streaming was enabled,
produced frames as seen through the incrementing frame counter.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417622
Original-Revision: 9553684f703afc23c66e499ac418a6d2eab4fa3e
GitOrigin-RevId: 899aabca427c6b6d65712e6a2b69da9b5107c1b7
Change-Id: Iabb3c6c3a438212943a927c8572de3cb56bbc7b1