[roll] Roll fuchsia [gidl] pin generated dart code to 2.8
This comment will tell the dart compile to compile this code using
version 2.8. This is required because this generated code does not
belong to a package in any package_config.json file so the compiler
will assume it should be compiled against the latest version. When
we turn on nullability this will fail to compile since it is not
converted to null-safety.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418277
Original-Revision: d7f5c1db3f0c6d18f82b3fe433b0d21bfc374549
GitOrigin-RevId: 44c9f0c28389b65f6cb9d2c1bec3bf86d6d8f6be
Change-Id: Idf4197f7d5e354a7bae03613ab0a47a05abb480b