[roll] Roll fuchsia [e2e] Also test for time zone *changes*
The dart runner should respond to time zone changes, as well as
time zone initialization.

This change modifies the timezone e2e test to try reading the timezone,
then change the timezone again, and ensure that the clocks match up
inside the dart VM and outside.

Previous test tried to do this only once, and didn't test repeated
attempts.  This works for the "deprecatedtimezone" implementation,
because it does not consult the time zone when computing the time
offset. But the "intl" implementation is sensitive to this, since it
asks about the timezone to compute its offset.

There are two tests for this feature; however, for the time being one
is `#[ignored]` because the two tests seem to step on each other while
changing the global mutable state (the time zone).

Multiply: timezone-test: 50
Original-Bug: 42387, 58383
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419497
Original-Revision: 114e7c60e8caa9429bafb55231ed113a50fe33a7
GitOrigin-RevId: 26cd7c5de573c1d9896bf2ed941171d4029d0b2f
Change-Id: I7341e34aa8fbf44c6440f3219f9e69cac8ac9728