[roll] Roll fuchsia [power_manager] Set up standalone Cobalt Metrics
The ThermalPolicy node currently contains all of the logic surrounding
Cobalt metrics reporting.  This has been fine so far, but became a
problem as soon as a second node was added that also contributes to the
Cobalt metrics "state" (that node is ThermalShutdown).

This change pulls the majority of the Cobalt metrics logic out of the
ThermalPolicy node and puts it into a standalone singleton instance. The
change also sets up the ThermalShutdown node to call into this new
Cobalt structure when the device reaches thermal shutdown.

Test: fx test power_manager_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416881
Original-Revision: 6f9a46fd64b8b2e7a0e7c4053a80e9d21093c306
GitOrigin-RevId: b078fe261ebb98f86a34df33e3c4873a7e1ee83c
Change-Id: Ic22788ea6d57a410bc92d80582aa29cab23fb60a