[roll] Roll fuchsia [kernel][x86][kasan] Check out-of-bounds accesses to globals
This CL instruments accesses to global variables in zircon.

The compiler adds a redzone to the right of every global object. Our
instrumentation poisons the redzones and in the steady-state traps on
accesses to the redzones to detect and panic on out of bounds accesses.

To enable this instrumentation, we add all kernel sections that can
have globals to the regions kASAN tracks; we also record every
contiguous region of globals in early boot and register them with
asan (posion their right redzones) in later boot.

Pair-programmed-with: mvanotti@
Original-Bug: 30033 Kernelmode ASAN for Zircon
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392703
Original-Revision: 12f9ce701550aed89c3c31efb0651e3533a0e7be
GitOrigin-RevId: 33d59e5ccf6bfba31c88e6419797d38eeaa1bccb
Change-Id: I122591ea1527f3c1bdc80c35feec0a87d3ebe06b