[roll] Roll fuchsia [camera][imx355] Connect the driver to the board
This CL establishes the connections between the hardware resources,
such as GPIOs and I2C interfaces, to the IMX355 camera sensor driver.
It also adds the imx355 to the device enumeration test.

Note that the EEPROM in the IMX355 camera modules will be connected
in a future CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417623
Original-Revision: 256514aaa03e47aeb8033ae6edc61c8879a93267
GitOrigin-RevId: f39eaa0d1ae9039366c9d1af96d7dd8b815336d1
Change-Id: I4c11c389272c63d4400a9c33d627fcc4f98e978f