[roll] Roll fuchsia [go] Add go.mod in subdirectories containing Go code
This is an alternative to having a single root go.mod, which doesn't
work due to the Go toolchain being built as part of the Fuchsia build
process. When the Go toolchain make.bash runs, there's eventually a check
to ensure that it is not inside a go module.

This is a bit tricky as:
- /garnet depends on /src.
- /src depends on /garnet and /tools.
- /tools depends on /src and /garnet.

A follow up will include "replace" statements to make this dependency tree work.
This can't easily be added now because a valid version containing a go.mod is
required.

To be extra clear: this does not affect the GN build at all, since all go.mod
are not in the root directory. But this does help the IDEs a lot, making things
like goimports just work.

Original-Bug: 57231
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/410950
Original-Revision: 3226fa91b98e44594f6094bb718c68210067f0a4
GitOrigin-RevId: e89a616921027f252f29e1578a1b790720c57826
Change-Id: Ic6369d43ee6bd5ba94aa644e7879204a92f736d1