[roll] Roll fuchsia [sdk] Rename core.api to core.manifest
The ".api" suffix on this file triggers an API-Review flag in the code
reivew tool, which is normally what we want for these golden files.
However, this specific golden file represents the contents of the IDK,
which shouldn't be gated by the API-Review flag. Instead, modifications
to this file are gated by Code-Review from the //sdk/OWNERS.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419495
Original-Revision: d2a01dae4fc31657de3a3e96b620814bef3a933f
GitOrigin-RevId: f2e3e38e34430b1d79313fd0eb9f8ee89c2a021f
Change-Id: I647955811f06c6e4585164247531fc47d869289e