[roll] Roll fuchsia [build] Support Zircon-specific toolchains in Fuchsia build
This CL introduces changes to the Fuchsia build system to
support toolchains that can build Zircon artifacts. In particular,
it introduces a new toolchain instance
(//zircon/bootloader:efi_${current_cpu}) which can be used to
build Win32 PE32+ executables for the Gigaboot bootloader.

Note that this CL does not try to build the bootloader with
the new toolchain, this is left for a future CL. Instead, it
introduces the necessary changes to the build system to support
this, and modifies compare_toolchains.py to verify that the
bootloader toolchains, as defined in the Zircon and Fuchsia
builds, are strictly equivalent.

A few notes regarding the choices being made:

- The ZN build uses the global 'toolchain' variable.
  This is implemented as 'zircon_toolchain' in the Fuchsia
  build, with a little twist: its value will be 'false'
  for non-Zircon toolchains, otherwise it will be a scope
  that duplicates the ZN build 'toolchain' variable.

- Also the ZN global variable 'is_gcc' is moved to the
  'zircon_toolchain' scope, since there is no plan to
  support GCC builds for non-Zircon artifacts.

- The zircon_toolchain() template is modeled around
  the c_toolchain() template found in the Zircon build
  at //zircon/public/gn/toolchain/c_toolchain.gni,
  but will generate a 'toolchain_variant' global
  variable that follows Fuchsia build conventions.

- For now, there are no attempts to match or
  re-implement the variant selection mechanism from
  the Zircon build. In particular its use of "tags"
  to filter variants during toolchain selection in
  variant_target() in //build/config/BUILDCONFIG.gn.

  This will be addressed in a future CL.

- Many of the configs introduced in
  //build/config/zircon/BUILD.gn have an equivalent
  in //build/config/BUILD.gn, but this CL doesn't
  try to replace them for simplicity. This will be
  performed in a future CL.

The next CL following this one will actually build
the bootloader with the new toolchain, and generate
a file that is byte-identical to the Zircon build,
except for some liberal GUID embedded by the linker
(similar to a build ID note for Win32 PE32+ binaries),
it is kept separate to ease the review of the current
CL.

Original-Bug: 54161
Testing: Run zircon/public/canaries/compare_toolchains.py
     to verify that the bootloader toolchains in the GN
     and ZN builds are identical.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401619
Original-Revision: cb3dafd2d16c0afe3ed7189ea168c87a4daba494
GitOrigin-RevId: f16266e719338c7aeaaba2b07a6dca4cb93ff85c
Change-Id: I998d5548de3147b310124f7cc3b640e8be2577d4