[roll] Roll fuchsia [storage] Explicitly zero fill in compressed FVM sparse images.
At the moment in FVM sparse images, there's no way to tell apart cases
where zero-fill is required and where we don't care. Since zeroes
compress really well, this change is explicit about zero filling and
indicates via a flag in the header that zero filling is then not
required. I've tested that this makes a neglible different to generated
size.

This should help a small amount with flash wear on devices that do a lot
of paving/flashing because we don't need to zero fill so much.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419536
Original-Revision: 08e4ee338fff0a3a6d6759fde93ffe47170ab8aa
GitOrigin-RevId: 503e583e923f6daae4c306d337a229cbd3e73bbe
Change-Id: I35b5172dcfbe95d248b923d5e88e862ba5da3f23