[roll] Roll fuchsia [camera][imx355] Fix crash on boot
This CL fixes a crash when booting a device that uses the imx355
driver. The root cause was the ISP query of the number of
available modes, which asked for more sensor modes than were available.

The fix was to change the logic to return the smaller of the
number of modes available and the number that was requested.

This CL also removes some spurious driver logging during driver
initialization.

Tested on a luis platform board with the camera in both the
muted and unmuted states at power on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419754
Original-Revision: 71090602cba58192a74131461cf7e970d0b3af0a
GitOrigin-RevId: 5343caab8a0a1641f2eaaede73f6cd4671608000
Change-Id: I88b49c4a2a4a20c1c2dd10e40d17c2d7553988cf