[roll] Roll fuchsia [audio][amlogic] Make sure codecs are muted when starting
This CL introduces an override_mute flag that is always checked
when updating the gain state in the codec(s). With this we can
guarantee that we'll always set the mute flag in the codecs before
we actually start a playback. Added tests to confirm we always
set muted in the codecs upon boot, and also to check the logic
on updating the gain state (including the added aggregation of
codecs mute and agc reporting).

Test: aml-g12-tdm-test
Test: Manual check of YouTube audio in Astro and Sherlock
Test: No pop on reboot on Astro and Sherlock
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418234
Original-Revision: b4639d180a91bac61f393c9980c82d00612958b7
GitOrigin-RevId: 1c45879b61e15c94fa176f5f0177edd559d599a4
Change-Id: If750c94f76b4b437a1274d6594d82c384aec97b2