[roll] Roll infra/recipes [tree_closer] Allow ignoring builds based on summary
The new summary_regexp_ignore property to the tree closer lets us ignore
builds with certain failure summary text when deciding whether to close
the tree. Initially we'll used this to ignore checkout failures -- if a
checkout fails, the build will have a summary like "Infra Failure:
Step('checkout.jiri update') (timeout)".

Ideally we'd be able to target step names, rather than summary text, but
that wouldn't work for fuchsia.py builders because the checkout happens
in the subbuild, so the only way to tell if the checkout failed based
solely on the orchestrator build is to inspect the build summary, which
the fuchsia.py recipe propagates from the subbuild.

Original-Bug: 56446
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/420197
Original-Revision: 67b818d941e375e8d4523394b361c8c555be9368
GitOrigin-RevId: b88653c6486d2eeb8884cb47049d2a52efd60bcd
Change-Id: If814455e6b02664e72f519ddc2f93dbc0b2863d6