[roll] Roll fuchsia [debugger] Give process name to empty module name.
By conventoion the dynamic loader sets the module name of the main
binary to empty. This produces output from "libs" and "sym-stat" which
is confusing to users.

This patch fills in the process name for empty module names. This isn't
always guaranteed to be perfect, but at least gives some indication of
what the binary is.

Original-Bug: 5233
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419922
Original-Revision: 6550a7b0bce7e2410a4f0e04b7cc6f20034562d7
GitOrigin-RevId: ab23d25e8ed5fa69e4d8aa445a777afe66957e84
Change-Id: Ib387bb9742ee0cb293c1bb17a9b57821696f8453