[roll] Roll fuchsia [drm] Add reply to GenerateLicenseRelease
When a DRM license is to be removed from a device, the
ContentDecryptionModule must first destroy the keys associated with the
license and then send a license release message that includes the record
of destruction. It is sometimes useful for clients to be notified of
these events distinctly so that they can update state accordingly. In
particular, it allows EME clients to resolve the
MediaKeysSession.remove promise once the keys have been destroyed and
before the LicenseMessage has been generated.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420238
Original-Revision: 9feafda26923c8ebd5f8ee144967a291fea61e22
GitOrigin-RevId: 5f7b9cc0a447ab6b20d2370ae00c71c78beb6562
Change-Id: I22a2a6b2e4e70de4f71d83532cad259462f053cb