[roll] Roll fuchsia [vulkan] Return an error when fence passed to vkAcquireNextImageKHR
The swapchain won't ever signal a fence passed to vkAcquireNextImageKHR,
and it's not possible to implement signaling that fence  - we don't
support external fences to signal it on the CPU, and no queue is passed
in that could be used to signal it from Vulkan.

There's an assert to protect against callers passing in a fence, but the
copy in the SDK is compiled without asserts enabled. So signal a debug
callback and return an error in that case.

Multiply: image-pipe-swapchain-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418955
Original-Revision: 01871c15dd68cadc7144120eb61ac4bf41ed1ae3
GitOrigin-RevId: 6c8c35022fcfbfc5d42eb0dd41c84c21bfb05b45
Change-Id: I25bc0464428e567c9038d8622fd306a1d7d8e958