[roll] Roll fuchsia [block][mbr] mbr boot signature incorrect logging
This fixes a bug in which the incorrect signature is being logged when
mbr.cc detects an invalid boot signature.  Currently, the code produces
a log that looks like:

  mbr: invalid mbr boot signature, expected 0xaa55 got 0xaa55

The wrong value is simply being used in the logging statement.  This
updates the correct boot_signature value.

Test: fx test -o mbr-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416875
Original-Revision: 931034411953b09cc33296cd451f3a7fac84db19
GitOrigin-RevId: 4a7161e329d17a80f47649a829763a3b720e7ec7
Change-Id: Ic5dbbd11d55de0bd55768e7c38fcbea2ccf8a38d