[roll] Roll fuchsia [triage] Separate error string for MISSING
This is a temporary solution until the "missing" refactor is done. For
now it is useful to know if there was a parse error or if something is
just missing.

TESTED=Manual fx triage. This is human-readable output only and should
not be asserted in tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420330
Original-Revision: 651b44ab664f9b15cd8a949c62692c92d65d62cc
GitOrigin-RevId: c7f2570f815d6d65339d9f19c0abdb0db55e3b71
Change-Id: Iedd4135b87543f57ddadbd321a7346992e82ad27