[roll] Roll fuchsia [a11y] Screen Reader unit test refactor.
This change refactors the Screen Reader unit test.

The unit test, as is right now, is not testing Screen Reader behavior, but the actions themselves. This makes hard to write new tests for the Screen Reader, because they need to replicate the action infrastructure.

This change makes the Screen Reader test its own behavior, mocking the result of the actions (which are already tested by the actions), and also make an abstraction to call the gestures (which are already tested in other places).

Original-Fixed: 49263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419478
Original-Revision: a4426bc3475437bbfc95944a7c19a62afbaea966
GitOrigin-RevId: 92465e1b09b716fb9de2c971257e08002df2f3b5
Change-Id: If8ada14a12d8c29f5a78e75dd3b87abb8e98fd06