[roll] Roll fuchsia [fx] speed up fx-update-path
I noticed that fx-update-path was responsible for almost half of my
shell startup time so I looked into what it was doing. It was calling
out to external processes a lot so I changed it to simply use builtins.

Much of it didn't actually work:
 - because the regexes to remove were actually the full paths changing
   your $FUCHSIA_DIR and re-running fx-update-path wouldn't actually
   update your path,
 - because tools/devshell/lib/vars.sh wasn't sourced fx-config-read
   wasn't defined so the Fuchsia and Zircon tools directorires weren't
   included in the path.

The core operation is in __update_path which takes a list of
$FUCHSIA_DIR-relative paths and ensures that the only members of $PATH
that end that way are the relative to the current $FUCHSIA_DIR. There
are separate implementations for Bash and Zsh because Zsh's scripting is
much more expressive.

From my tests this reduced the runtime of fx-update-path from about
110ms to about 50ms in bash and about 115ms to under 10ms in zsh.

Test: fx self-test fx-internal/fx_update_path_test
Test: fx test fx_script_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416415
Original-Revision: a71c2782efdd16c2eacdb3d01f552eb2c2244b1b
GitOrigin-RevId: ef1cb66e62de88f9386c372fcad2076197dc2175
Change-Id: Ifecb720ef97691d10c8d5977f3a56b2777c2e583