[roll] Roll fuchsia [kernel][vm] Explict page list node ranges
Modifies the VmPageListNode to have separate ForEveryPage and
ForEveryPageInRange methods, with the latter requiring that it be given
a range that is strictly inside the nodes range.

The VmPageList::ForEveryPageInRange is then modified to iterate the
range and know whether it is operating on a partial node range or a
full node range instead of throwing ranges at nodes and seeing what
sticks.

Motivation for this change is that knowing and being explicitly in the
range being operated on allows for future refactors and improvements.

Test: Covered by existing tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416539
Original-Revision: 20ca084f440690b094df1d6b65ba96c4ae1fd785
GitOrigin-RevId: d31c8bd853ed40c400d9bdf430ec58566cab89ee
Change-Id: I9f3e0170cd47882653e2036fdb9a51fad78acd69