[roll] Roll fuchsia [media][audio_core] Bringup AudioCore per-test
Currently, a single instance of AudioCore is used per test suite; this
enables state to be carried between test instances within a test suite,
which causes flakiness if test order is shuffled.

This CL enables an HermeticAudioEnvironment to be created per-test
within a test suite, such that each test has a unique instance of
AudioCore and state is not shared between test instances.

A slight decrease in performance is found, due to the environment being
recreated with each test instance, but the stability of the tests
outweighs this performance decrease.

Performance of:
    fx test audio-core-api-tests audio-core-api-pipeline-tests

with HermeticAudioEnvironment per:
    > test suite
        real    0m45.324s
        user    0m7.671s
        sys     0m3.123s

    > test instance
        real    0m51.228s
        user    0m8.385s
        sys     0m3.631s

Original-Fixed: 56242
Original-Fixed: 56243
Original-Fixed: 56247

Test: fx test audio_core_unittests audio-core-api-tests
audio-core-api-pipeline-tests -- --gtest_shuffle

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418756
Original-Revision: bebf7556dd267aaeb965346de1b12eac8694c6ed
GitOrigin-RevId: 9962dfd9827153062f55f4d69585dc8f64fda637
Change-Id: I4c483194b0c760b17cea2ea9eecaaa463137a4fc