[roll] Roll fuchsia [kernel][vm] Standardize on pointer for callbacks
Some VmPageList callbacks used a reference for the markers, and some
used a pointer. This caused inconsistencies in usage and required
wrapping lambdas to convert to the internal methods, that all used
pointers. Given that some methods are non-const and allow modifications
standardize on the pointer variant and remove unneeded wrappers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420078
Original-Revision: 2262c46375912ebc4b0466de468bd8bc4346b44c
GitOrigin-RevId: 1aab3b602fa2ff6f467545fa2fe1e915dfd49300
Change-Id: I4350d5e759a218088bf9623bba5a9a4b6f816bcf