[roll] Roll fuchsia [fidl] Add check() to QueryResponseFut
Adds check() to QueryResponseFut, allowing an early check for the
message having been sent.  This is similar to the check that is
performed by one-way messages that have no response.  If an error
occurred in sending, it is returned, otherwise a QueryResponseFut is
returned that can be polled for completion.

Test: added client_query_result_check
      updated other tests for client_raw_query API change

Multiply: fidl-rust-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418766
Original-Revision: aff56e40caef0fcbd18d500391062f25416b4e61
GitOrigin-RevId: 5c39dc48dba078e46a705f30870a2f0409986998
Change-Id: I9a295671c613560036a61e87925f8c3b6e654253