[roll] Roll fuchsia [retry] Fix flake
We started seeing flakes in this test starting on August 18. The flakes
happened because if multiple channels are ready, Go's select statement
randomly chooses one of them to read from. This was problematic for the
select statement in `retry.Retry()`:

  timer := time.NewTimer(next)
  select {
  case <-ctx.Done():
      timer.Stop()
      return err
  case <-timer.C:
      if c != nil {
	  c <- err
      }
  }

When `next` is zero (e.g. when using `ZeroBackoff`, like in the flaky
test), even after the context is canceled `select` won't always enter
the `<-ctx.Done()` case due to the aforementioned nondeterminism. We can
fix this by checking if the context is canceled before entering the
select statement, to ensure that we'll always exit on the first
iteration after the context is canceled.

The code has been this way for a while with no flakes, so I suspect this
started happening because of the recent upgrade to Go 1.15. The
nondeterminism of `select` isn't anything new, and I can't find anything
related to it in the changelog, so it may have been caused by some small
internal timing-related change.

I also improved the error messages in the tests and fixed a timer leak
in the Retry function.

Multiply: retry_tests: 1000
Original-Bug: 58651
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420714
Original-Revision: 4807b3f9472518dff5b786fce0954593d13a56af
GitOrigin-RevId: fa6d3fa48e8dcc01df2a38d204b2bd9be902c64f
Change-Id: Ia87b03ec771d07f6566a825888f7eb4a8a7b7624