[roll] Roll fuchsia [syscalls][kazoo] Map bool in syscall fidl to uint32_t in C
This is necessary because the compiler is allowed to assume that the
incoming bool is either 0 or 1 on the incoming side, resulting in
incorrect code generation (but with no way for the kernel side to
actually validate the value). By making it an explicit integer type, the
kernel side can range-check.

This causes a change from bool to uint32 in C, Rust, and Go bindings,
however there are currently no bools used in syscall fidl after
https://fuchsia-review.googlesource.com/c/fuchsia/+/418836. As a result
there is no (further) change to the syscall api exposed by the system.

Original-Bug: 53118
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419682
Original-Revision: 18bd75e7e939df0551ed0d27f4fad606cdc21d3b
GitOrigin-RevId: aa34e09532e156520033e3ab5e5e1c561d26f785
Change-Id: Iddf6d3797bc22a9959f8a69585676b1eac62ed90