[roll] Roll fuchsia [build] Support variants in compare_toolchains.py
This modifies the implementation of the compare_toolchains.py script
during build unification to support variants as well. This feature
will be used in a future CL that adds toolchain for the multiboot
Zircon targets, which can be built with the `gcc` variant as well.

Original-Bug: 54160
Testing: Running the script [1] should still succeed.

[1] zircon/public/canaries/compare_toolchains.py --clean /tmp/results

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417840
Original-Revision: 881275753f72556695f88f3d103c677f339f297d
GitOrigin-RevId: 55472d2dc922b42fdcfc2adb593ffbe2a2e628d7
Change-Id: I1f99134a3a34a32cd744e7245cdefb82bd865bca