[roll] Roll fuchsia [resultdb] Update logic for setting the "expected" field.
We have tests that fails expectedly (ex:
host_x64/zxtest-integration-test). We are storing the result correctly
in summary.json where the testSuite reports the test as "PASSED", and
the testCase reports tests as PASSED/FAILED based on what really
happened from stdout.

We want to record the "expected" field in ResultDB to indicate the
correct test behavior. If a "FAIL" is correct, then "expected" field
will be set to True; and if a "FAIL" is incorrect (99% of the time), the
"expected" field will be set to False.

Original-Bug: 58744
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419736
Original-Revision: 098ad675ed0a419ad645138a8a2d0700b8f9fb12
GitOrigin-RevId: d65ea09e4846a112ea0702510cf55385c1707622
Change-Id: Ibd7e046664c0bebf42ced8f75494985993b437c6