[roll] Roll fuchsia [blobfs] Fix flakiness in AllocateIncrementsMetricTest
The inspect reader service was incorrectly making the first recursive
call to the Tree FIDL client immediately on the calling thread and
successive calls would be wrapped in a promise that would run on the
executor, possibly in another thread.

This is bad behavior because the FIDL client is thread-hostile
and caused flaky behavior every 5000 runs or so.

This CL makes the following changes:
* ensures that inspect::SnapshotTreeFromTree calls are wrapped
  in a promise that will run only on the expected thread.
* in AllocateIncrementsMetricTest, quits the loop and joins on
  the thread before destroying the Executor. This ensures there
  are no running tasks when the Executor gets destroyed.

Original-Bug: 55829
Test: `fx shell runtests -n blobfs-integration-test -- -i 1000`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421114
Original-Revision: a3f2eeda48f09c1082b4be8731af879e7ad9aa46
GitOrigin-RevId: 602aec58e1725ea1cdb3a5e7b3efc6cb773c93ec
Change-Id: I6e5daf4b8caec08ecd471284f272600ea139d5e7