[roll] Roll cobalt [Cobalt 1.1] Update event_code_buffer_max field of MetricDefinition
Updates comments to clarify that the field is optional and
will default to the total number of event vectors for all
MetricDefinition types. Since the number of event vectors may
be a 64-bit int, the type of the field is now uint64.

Also slightly changes the meaning of the field: rather than the
number of distinct event vectors which the Cobalt client tracks
at any given time, it is now the max number of distinct event
vectors which should be included in any single observation.

Original-Bug: 58854
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/420857
Original-Revision: 23043aea43c1189ad6be76c6148ea3f6c7ad5a4d
GitOrigin-RevId: 402281c7178b6081ee0d5dfaf5b593c6dabf7bb5
Change-Id: Ia3deb5b4e07474cbc19a1a1816ed99f752f2fae6