[roll] Roll fuchsia [ui-input][root_presenter] Add multiple views
In preparation for converting RootPresenter to using the new
fuchsia.ui.pointerinjector API for input injection.
Inserts two extra views internal to RootPresenter to use as context and
target for injection.

Adjusts transform logic to make this easier:
The scene and the root view is now matched to the native rotation and
resolution of the display.
The internal child view (which will eventually be the
"accessibility view" fxb/50549) is rotated and scaled to match the pips
instead.
This allows input to be injected raw and have scenic handle any
necessary transformations.

Original-Bug: 49819
Manually tested on emu, astro and sherlock.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407728
Original-Revision: 3bb27b5f560622a654dae107410993f26dccbfd9
GitOrigin-RevId: 120d43bfc25d4d22bb13e72528caf6246be710fc
Change-Id: Iad33abbb13741efa892e31ca48c5365a893ce9c6