[roll] Roll fuchsia [audio][test] Fix min lead time race
In some tests, when starting a renderer, the following event sequence is
rare but possible:

1. Create renderer
2. Ask for MinLeadTime events
3. OnMinLeadTimeChanged(0)
4. Renderer links to output
5. OnMinLeadTimeChanged(>0)

We are using the first OnMinLeadTimeChanged event as a signal that the
renderer has been linked to the output. Per the above sequence, we
should actually wait until the lead time is non-zero.

Original-Fixed: 58394
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420817
Original-Revision: 97f846097de73d2970780f1952b32c2327ecbe17
GitOrigin-RevId: 335c0bf9f35646e6f2e98b270855f3687dab0c7f
Change-Id: Ib49db74d6046d99a85945e86ad1cb91c067ef4cd