[roll] Roll fuchsia [SL4F] Add Reading DDR windowing data
Allows reading a 'sticky' register which stores
DDR windowing tool results that were generated
by the bootloader.

Included in the CL are the changes to the fidl API,
the implementation in the driver, and the facade
which reads the value for factory testing.

Test: Facade tests are included. A follow-up CL
allows manually testing the driver via the CLI.
An integration test of the facade is in a followup CL
in the factory repo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419317
Original-Revision: 2169e0c6531118b60cfc113b74ddf53a43c4675c
GitOrigin-RevId: 82ef952c88b025aecf0b7e700efb318ea171cd01
Change-Id: Iee8f4df8165e1988af8c810f97bd05a9771a2438