[roll] Roll infra/recipes [lsc_check] Add recipe to validate LSC footer
This is to make sure that stakeholders are notified appropriately, via a
bug, of large-scale fuchsia platform changes that affect other projects.
For now, this recipe just checks that the footer exists on changes to
certain files. Eventually we might want to actually get details about
the bug linked in the footer and/or file a bug for the CL owner if the
footer isn't present.

Original-Bug: b/166160654
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/421178
Original-Revision: 2210c1f7fd7fc914db04f553d66d681b35d6535e
GitOrigin-RevId: d1617f370a79692450643b0cff24b405c691d417
Change-Id: Ic8bc3848921d2a69e8ad844dc1f6ef73db5598b7