[roll] Roll infra/recipes [release] Fix get_next_candidate_version for rolled-back branches
Conceivably, a branch's history can look like

0.20200817.1.2
0.20200826.1.1
0.20200826.0.1

Where the tip of the branch has been rolled back to a version older than
version(s) in its history.

In the case that get_next_candidate_version is called, e.g. during a
cherry-pick, the function will incorrectly compute the next candidate
version to be the max() of the release versions on the branch, + the
candidate version incremented == 0.20200826.1.2.

The true version should actually be 0.20200817.1.3.

This change fixes the function so that get_next_candiate_version simply
considers the ref that is passed into the function as the base release
version to increment.

Original-Bug: 59017
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/421340
Original-Revision: a5b0fa58beae6583b4319ec3b36d0d3eb42116ac
GitOrigin-RevId: 9ab1f2609112b5c9d81f2434adfb0f3f62157343
Change-Id: I8a12a104f9f3eb0a2e6f83aacfb9f0e1a84330be