[roll] Roll fuchsia [kernel][vm] VMO hierarchy generation counts
Add a generation count associated with each VmObjectPaged hierarchy,
which is incremented on any operation that could result in a change in
page attribution counts anywhere in the hierarchy. The generation count
is used to implement caching of page attribution counts (changes for
this to follow in a separate commit).

This commit just introduces the generation count, and increments it in
the required places. The generation count is not used anywhere yet.

These are the spots where the generation count is updated.
  1. Creation and removal of clones and slices.
  2. Committing a new page in GetPageLocked().
  3. Resizing the VMO, decommitting or zero'ing a range.
  4. Taking and supplying pages via supply_pages().
  5. Evicting a page, or deduping a zero page.

Original-Bug: 56779
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417454
Original-Revision: 0c26b48ae077f07ed090b7225ed21f6875253b3b
GitOrigin-RevId: f7fe6c7dc626405c137a79a7ad283c95e9cbb904
Change-Id: I9f8eb7b83cd3853595c88eecc123f2025163de79