[roll] Roll fuchsia [amlogic_decoder] Heuristic for lower latency decode
Since we don't presently get max_num_reorder_frames
from the FW, for now we can lean on this heuristic to
output decoded frames sooner.

This heuristic isn't perfect.  It's possible to make
an h264 stream that causes this heuristic to output
frames in an incorrect order, but not many frames,
and any potential out-of-order output is only near the
start of a stream.

We can disable this if/when the FW starts providing
max_num_reorder_frames.

For clear streams we could directly parse the SPS to
get max_num_reorder_frames, but the downsides of that
seem larger than the risk of this heuristic.

BUG: 58101
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420976
Original-Revision: 5be111106d85741bdfc61ed9ce2ddc981e1c7fe8
GitOrigin-RevId: 078265cd6d519f373a812dff158247cd438418cf
Change-Id: Ieca5c942ad78b20fe2d601df49285f244beda765