[roll] Roll infra/recipes [git] Add checkout_from_build_input method
The git recipe module provides `checkout_commit()` and `checkout_cl()`
helper functions for initiating a checkout based on the Buildbucket
input. However, this leaves it up to the caller to determine which
function to call based on whether
`api.buildbucket.build.input.gerrit_changes` is empty, which is annoying
and leads to a lot of duplicated code. So we address that by moving that
logic into a `checkout_from_build_input()` function that inspects the
buildbucket input to determine whether to call `checkout_commit()` or
`checkout_cl()`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/420460
Original-Revision: b973138116ff41ade44e9a849c50bc4b71799d74
GitOrigin-RevId: a6350458f66e5db53ed35b2112c73e873eab23b1
Change-Id: I5f2fa661fec7cf0a09e1538da092ea27d9e60a0e