[roll] Roll fuchsia [watch_handler] Support watches where initial value is not known
This extends the watch_handler library to handle cases where the
initial value is not known at the time the WatchHandler is created.
This can occur in cases when producing the initial value is also an
asynchronous operation that may not be complete by the time the first
Watch client connects.

Testing: fx test watch-handler-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418755
Original-Revision: 03d35c2acfbf9a1c21d0c7457c464122d2452210
GitOrigin-RevId: bd718ec63e5c8b763ffeea37a50aad66616c657b
Change-Id: I756148170157cab81edbb2adcc277cff6d63135d