[roll] Roll fuchsia [time] Remove network_time integration test.
This test ensured the correct behavior of the original time architecture
where network time service was the coordinator, and won't work as we go
through the transition to the intended arhitecture where timekeeper is
the coordinator. Per the time test strategy we will eventually have a
timekeeper integration test offering similar coverage, but in the
interim we will rely on unittest and the time e2e test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420721
Original-Revision: 627c41a1f12fe9fc51f11da130722d86f7851562
GitOrigin-RevId: 50c229264a74d0b42005c55b826ec306c50c6e8f
Change-Id: If5c9cd95a117cccddb70c94021755e412411ef66