[roll] Roll fuchsia [power_manager][luis] Use ThermalShutdown node
The thermal policy on Luis is currently very simple and is only required
to shutdown the device at a configured temperature threshold.

The policy was initially set up to achieve this by using the existing
ThermalPolicy node. Now, we have the ThermalShutdown node which is a
dedicated node for this purpose.

We should use the ThermalShutdown node until we have a reason to use the
ThermalPolicy node on Luis. This will let us be more flexible in driver
transitions on Luis. Specifically, it lets us remove the CPU driver
(temporarily) on Luis without breaking power_manager dependencies.

Test: fx test power_manager_tests
Test: Boot up a Luis board with the change and verify the thermal
  shutdown logic behaves correctly by lowering the reboot temperature
  threshold and then artificially increasing the device temperature and
  observing a system reboot.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420742
Original-Revision: 41f480c9134afbfef172bd365c205772ad553d81
GitOrigin-RevId: 402f27e6c39ca97d3db7257b429d6daa81fc66de
Change-Id: I224f24b18d6705ca3892d84b190cd712ba93f3c5