[roll] Roll fuchsia [aml][rawnand] Remove redundant irq thread
The aml-rawnand driver uses a dedicated thread to wait for interrupts.
This thread does not do much except wait for the interrupt in a while
loop, and signal a sync_completion_t when the interrupt arrives. The
sync_completion_t is only waited on from AmlQueueRB().

This change gets rid of the irq thread and the sync_completion_t.
Instead, we now wait on the interrupt object directly from AmlQueueRB().
This saves the cost of the extra context switch, and also gets rid of
the timeout of 20s in AmlQueueRB() to wait on the sync_completion_t.

This is purely an optimization, and does not change the functionality of
the rawnand driver.

Test: Verified that the device boots and is functional. Verified
presence of read interrupts (0x42) when nand-worker thread is running
from traces.

Original-Bug: 57899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421815
Original-Revision: f966a6f83be18ac3080ecbf374bd5e712141a1c1
GitOrigin-RevId: ff49a26833a4368bc161bb19a51bc617c28842be
Change-Id: I08a4e9a49267ae57992fd02c096be932c82b80d5