[roll] Roll fuchsia [thermal][luis] Support aml-thermal-s905d2g
This change supports the migration from aml-thermal-s905d2g-legacy to
aml-thermal-s905d2g on Luis. It is the first change in a series to
achieve a soft migration:
  1. Add thermal devices to support both drivers on Luis
  2. Stop building aml-thermal-s905d2g-legacy on Luis; instead build
     aml-thermal-s905d2g
  3. Remove the thermal device that supports aml-thermal-s905d2g-legacy
     on Luis

With this change, both a platform device (to support
aml-thermal-s905d2g) and a composite device (to support
aml-thermal-s905d2g-legacy) will be added. The driver that ultimately
gets matched and loaded will be determined by the build configuration
(currently it is aml-thermal-s905d2g-legacy, but the next change in this
series will update it to be aml-thermal-s905d2g). Only one of the two
drivers should be built into the image for this reason.

Test: boot up Luis board with this change and verify power_manager is
  still successfully reading from the thermal driver
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420743
Original-Revision: ba1ea398f081897fa28cbd324e52214b67f225d3
GitOrigin-RevId: a1621eb5cfc7232d74093cfb62ecdad1eaa53dd1
Change-Id: I7b2d811c85ea7e72196a687a0202062a4be774ff