[roll] Roll fuchsia [fidl][llcpp] optimize sync event handler
The callbacks for handling events are modified. Now we only give a
pointer to the primary object. This way we avoid unnecessary moves.

HandleEvents now returns a fidl::Result

HandleEvents now uses a reference to EventHandlers and uses
fit::function instead of fit::callback. That means that users can now
create the handlers only once and use them for each invocation.

SyncClient::HandleEvents is now inlined.

That reduces the size of the generated code. That also reduces the size
of the user code.

size reduction of LLCPP Domain Objects: 6836 bytes
global size reduction (LLCPP + user code): 16384 bytes

The goal for this CL and the upcomming CLs is to only have callbacks
with one argument (a pointer to the primary object). This way, all the
callbacks will be consistent and easier to use. Most of all, by removing
all the moves (for the arguments), we are able to reduce the code size.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421236
Original-Revision: 360df6bc40a073d4c53c46e542898dab5e16e2e1
GitOrigin-RevId: a12f1de8f2abfb0bfd195096804e010cc3dbdc4e
Change-Id: Ie71a7540ae25dbc5588042e09c73af70782edb73