[roll] Roll fuchsia [sy8827][luis] Pass i2c params to sy8827 bucks.
There are two of these on the system and we need to be
able to distinguish between them somehow.

Pass the i2c address of the bucks as an argument so that they
can set it as a device property when they add a device.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421800
Original-Revision: e72d7fdd283156c76f20e34c4e403d989853b3fe
GitOrigin-RevId: 552b69b63ba773014bd18bc3ca7b7dea7f451618
Change-Id: Ibc352c369f78eb2446dd6d1b68f2dbd39a982faa