[roll] Roll fuchsia [bootserver_old] Start improving tests
It is not enough to fix the test, but it gets farther. In theory one should be
able to run it after doing:

    sudo tunctl -t qemu -u $USER

but I fail to run it in either "fx test" or "go test" mode.

- Rename bootserver to bootservertest as this customary in Go to suffix
  'test' to test-only Go packages.
- Add -test_data_dir flag, which will unblock running as go test later.
- Leverage t.Cleanup().
- Check more error codes. Leverage new error return value from qemu package.
- Reduce the overwhelming amount of whitespace in BUILD.gn. It looks like
  Kennedy for GN.
- Discover the output directory (out/default) based on .fx-build-dir.

The flag -test_data_dir serves to cope with the difference between "go test" and
"fx test".

- In "go test" mode:
  - Current working directory is the Go package's directory
  - Executable is in $TMPDIR
- In "fx test" mode:
  - Current working directory is the build output directory (out/default)
  - Executable is in out/default/host_x64 (or host_arm64).

The Go test historically did path walking based on the executable's location.
That doesn't work in "go test" mode. Hence I'm changing the path to the test
data to be explicit.

In "go test", the relative directory is known in advance, so I can hard coded it
as a default.
In "fx test", the flag can be specified in BUILD.gn, so I do that.

Original-Bug: 55387
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420517
Original-Revision: 622d20265d9e4448fb911bf04c824b3cf4657e88
GitOrigin-RevId: 595fb5a9f3979f1014b48170947eac1a007a9304
Change-Id: I8e5e78630380ef36c015ad683854656d6f288fc5