[roll] Roll fuchsia [components] *Actually* monitor for process exit
In 358e2f304771e5ca0c246ef6b5ba9dd45b734c06 I claimed to fix
this issue, it was not true, I accidentally left out part of
the change.

This change updates behavior so that for critical components
we wait until the process exits to clean up its job and mark
it stopped. For components that aren't critical the behavior
is unchanged.

Previously we waited on the Lifecycle channel to close before
ElfComponent's stop future returned. This could lead to a
condition where we would kill the component if the component
closed its Lifecycle channel before closing. This could be
a significant problem for components whose main process is
marked as critical.

The tests are also updated to prevent process leakage in the
case of panic.

Original-Fixed: 56794

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416828
Original-Revision: daf50b71ab34ab9412c68779ea06015390c7dced
GitOrigin-RevId: 2a19e40527f4d4f5619557ce0b886c167e8b3aac
Change-Id: I36a00c451768ac7307d322503ba4dbb57c3a7df9