[roll] Roll fuchsia [rust][build] Add rustc_version_string as build arg
This allows forcing recompilation when using a custom build of the Rust
toolchain.

We attempt to do the right thing in all cases: Using the CIPD instance
ID when a prebuilt is used, and falling back to the build arg (defaulted
to empty string) otherwise.

The Zircon build always uses the value derived in the Fuchsia build.
However, for cases where the Zircon build is being run directly, we
duplicate the logic to find the CIPD instance ID.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421735
Original-Revision: 7de1ebd6c45fe6fe65889f83fbf39837d04d9136
GitOrigin-RevId: c00ec07602685660c3746d218d9144319e9b5002
Change-Id: I6e6066383953403a6632cb8becb428ebe4e34d28