[roll] Roll fuchsia [block-verity] Extract VerifiedVolumeClient
This is a refactor that extracts a reusable `VerifiedVolumeClient` from
the existing integration tests for us to use in fshost.  The tests have
been modified to make full usage of `VerifiedVolumeClient`, which provides
test coverage of the client library.

Original-Bug: 58500
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422254
Original-Revision: f5928e3c2509de8b35a0924e2f03182630227e26
GitOrigin-RevId: 5adab79a69527277b15851821875fea8bf0419ce
Change-Id: Ic29832cf78f887d5ba3fb49dad363ddd4f5d17eb