[roll] Roll topaz [fidl][benchmarks] Add zx package as dependency in topaz repo
In CL, I7de61c3323940d03ed7db78895700ff90e4166f6 we add a FIDL
file to the benchmark set that uses the zx package.

In order for tests to pass in that CL, the dart benchmarks in
the topaz library need to first depend on the zx library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/topaz/+/422598
Original-Revision: bf775fabc3bcde00f388cabce557142358f70221
GitOrigin-RevId: 76459ece0459526f8c651f5b3b483e114f384a01
Change-Id: If7cc743d22f3aa025e45f2433f8da21040a25da0