[roll] Roll cobalt [cobalt 1.1 privacy] Look up privacy encoding params for reports
Defines a type PrivacyEncodingParamsCalculator which maintains a
lookup table of privacy encoding parameters.

The keys of the table are of the form

(epsilon, population, sparsity)

where:
* epsilon is the target (shuffled) epsilon value,
* population is the estimated size of the fleet,
* sparsity is the maximum number of buckets with nonzero counts in
  the histogram contributed by an individual client.

A key is a function of a MetricDefinition and ReportDefinition,
together with an estimate of the fleet size and a mapping from
Cobalt 1.1 PrivacyLevels to target shuffled epsilons.

The values of the table are of the form

(prob_bit_flip, num_index_points)

and a value can be used to populate the fields of the same name
in the ReportDefinition which was used to compute the key.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/418055
Original-Revision: 3f97f9b47e91f3a58ed12f1f1ddcdc999f1a8862
GitOrigin-RevId: e1713c3535c324c8cd477a7ab46968a1016c9712
Change-Id: I741e89733c16005464c9e8684b7d1ba5b8cff09b